// src/tenable/types.ts
export interface RecentScanSummary {
  id: number;
  legacy: boolean;
  permissions: number;
  type: string;
  read: boolean;
  last_modification_date: number;
  creation_date: number;
  status: ScanStatus;
  uuid: string;
  shared: boolean;
  user_permissions: number;
  owner: string;
  schedule_uuid: string;
  timezone?: string;
  rrules?: string;
  starttime?: string;
  enabled: boolean;
  control: boolean;
  name: string;
}

export enum ScanStatus {
  Completed = 'completed',
  Aborted = 'aborted',
  Empty = 'empty',
  Imported = 'imported',
  Pending = 'pending',
  Running = 'running',
  Resuming = 'resuming',
  Canceling = 'canceling',
  Canceled = 'canceled',
  Pausing = 'pausing',
  Paused = 'paused',
  Stopping = 'stopping',
  Stopped = 'stopped',
}
// ...
export interface ScanHostVulnerability {
  count: number;
  plugin_family: string;
  plugin_id: number;
  plugin_name: string;
  severity: number;
  host_id: number;
  hostname: string;
}
// ...
export enum VulnerabilitySeverity {
  Info = 'info',
  Low = 'low',
  Medium = 'medium',
  High = 'high',
  Critical = 'critical',
}

export enum VulnerabilityState {
  Open = 'open',
  Reopened = 'reopened',
  Fixed = 'fixed',
}
// ...
export interface ExportVulnerabilitiesFilter {
  cidr_range?: string;
  first_found?: number;
  last_found?: number;
  last_fixed?: number;
  plugin_family?: string[];
  plugin_id?: number[];
  network_id?: string;
  severity?: VulnerabilitySeverity[];
  since?: number;
  state?: VulnerabilityState[];
  vpr_score?: {
    eq?: number[];
    neq?: number[];
    gt?: number;
    gte?: number;
    lt?: number;
    lte?: number;
  };
}

export interface ExportVulnerabilitiesOptions {
  num_assets: number;
  include_unlicensed?: boolean;
  filters?: ExportVulnerabilitiesFilter; // See filter notes above
}
// ...
export interface ExportAssetsOptions {
  chunk_size: number;
  include_unlicensed?: boolean;
  filters?: ExportAssetsFilter;
}

export interface ExportAssetsFilter {
  created_at?: number;
  updated_at?: number;
  terminated_at?: number;
  deleted_at?: number;
  is_terminated?: boolean;
  is_deleted?: boolean;
  is_licensed?: boolean;
  first_scan_time?: number;
  last_authenticated_scan_time?: number;
  last_assessed?: number;
  servicenow_sysid?: boolean;
  sources?: string[];
  has_plugin_results?: boolean;
  network_id?: string;
}